Register Command, MongoDB Connection and Schemas, Logger, Input Handler
- Register Command
Will register, update, remove and list all the members stored in a MongoDB Collection.
- Database (dj.js)
Module used to connect and offers the application required and different connection methods to the remote MongoDB.
- InputHandler
Future CLI handler for adminitration and debug purposes.
- Logger Module
Centralized module to be responsible of logging each log to its specific file.
Still logging to FS.